trashbox-restore now returns and error on failure

+++ trashbox.sh
@@ -2,7 +2,7 @@ export TRASHBOX="${XDG_DATA_HOME:-$HOME}/.trashbox"
 
 # moves 1 or more files or directories to the trashbox
 function trashbox-put() {
-  if [ "$#" -eq 0 ] ;then
+  if [ "$#" -eq 0 ]; then
     echo "Usage: trashbox-put <path ...>" 1>&2
     return 1
   fi
@@ -15,18 +15,21 @@ function trashbox-put() {
 
 # restores a file or directory from the trashbox
 function trashbox-restore() {
-  if [ "$#" -ne 2 ] ;then
+  if [ "$#" -ne 2 ]; then
     echo "Usage: trashbox-restore <source relative to the trashbox> <destination>" 1>&2
     return 1
   fi
 
+  if [ ! -e "$TRASHBOX/$1" ]; then
+    return 1
+  fi
   mv "$TRASHBOX/$1" "$2" &&
   (rmdir "$TRASHBOX/$(echo "$1" | cut -d / -f 1)" 2> /dev/null || true)
 }
 
 # prompts to permanently remove everything from the trashbox
 function trashbox-empty() {
-  if [ "$#" -ne 0 ] ;then
+  if [ "$#" -ne 0 ]; then
     echo "Usage: trashbox-empty" 1>&2
     return 1
   fi
@@ -34,7 +37,7 @@ function trashbox-empty() {
   printf "Empty trashbox? (y/N) "
 
   read answer
-  if [ "$answer" != "${answer#[Yy]}" ] ;then
+  if [ "$answer" != "${answer#[Yy]}" ]; then
     rm -rf "$TRASHBOX"
     mkdir -p "$TRASHBOX"
   fi
@@ -42,14 +45,14 @@ function trashbox-empty() {
 
 # prompts to permanently remove some files or directories from the trashbox
 function trashbox-remove() {
-  if [ "$#" -eq 0 ] ;then
+  if [ "$#" -eq 0 ]; then
     echo "Usage: trashbox-remove <path relative to the trashbox>" 1>&2
     return 1
   fi
   printf "Delete the given files? (y/N) "
 
   read answer
-  if [ "$answer" != "${answer#[Yy]}" ] ;then
+  if [ "$answer" != "${answer#[Yy]}" ]; then
     for file in "$@";do
       rm -rf "$TRASHBOX/$1"
       mkdir -p "$TRASHBOX" # in case file is '/'
@@ -60,7 +63,7 @@ function trashbox-remove() {
 
 # lists all files and directories in the trashbox
 function trashbox-list() {
-  if [ "$#" -ne 0 ] ;then
+  if [ "$#" -ne 0 ]; then
     echo "Usage: trashbox-list" 1>&2
     return 1
   fi
